Almost remove buffering from core.Stream
The implementation of Stream now doesn't try to buffer data from the
underlying iterator until it has enough.  To get the old behavior, it is
enough to wrap it in a BufferedReader, as is done in
test_core.TestBufferedStream.  However, it still has an internal buffer
because the length of chunks from the iterator isn't known in advance
and, if larger that what is requested, needs to be stored internally
until at least the next request.