Harden source and tests based on mutation testing
The implementation and the tests were hardened so that Mutmut doesn't
find any mutant that passes the tests.

The default chunk size was removed so that it can't be mutated.  It's OK
since it was probably unnecessary (the user should decide what chunk
size to use).

The code was also reformatted with Black although that's not checked in
CI yet.